+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@ Note that this file was created and maintained starting with the `netcdf-cxx4 4.
 
 ## netCDF-CXX4 v4.3.1 released TBD
 
+* [Bug Fix] Added `ncFile::create()`, also added a new `open` function and constructor to allow for more flexibility when opening a file.  See [GitHub #55](https://github.com/Unidata/netcdf-cxx4/issues/55) for more information.
 * [Enhancement] Addressed an issue reported with `Intel Compilers 17.0.0`. See [GitHub #41](https://github.com/Unidata/netcdf-cxx4/issues/41) for more information.
 * [Enhancement] Updated how `configure` determines information about the existing netCDF-C install.  See [GitHub pull request #39](https://github.com/Unidata/netcdf-cxx4/pull/39) for more information.
 * Corrected an issue where cmake-based builds weren't generating `ncxx4-config`.  See [GitHub pull request #37](https://github.com/Unidata/netcdf-cxx4/pull/37) for more information.